backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.12
geoalchemy2==0.13.0
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.24.0
asyncpg==0.27.0
alembic==1.10.4
pytest==7.3.1
python-dotenv==1.0.0
websockets==11.0.3
redis==4.5.5
celery==5.2.7
